pyproject.toml
[build-system]
requires = ["flit_core >=3.11,<4"]
build-backend = "flit_core.buildapi"

[tool.flit.sdist]
exclude = [".*/"]

[project]
name = "gradupe"
authors = [{ name = "David", email = "wavim.wu@gmail.com" }]
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
dynamic = ["version", "description"]
keywords = ["image-dedupe", "sobel-gradient"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Typing :: Typed",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research"
]
requires-python = ">=3.10"
dependencies = [
    "imagesize>=1.4.1",
    "numba>=0.62.0",
    "numpy>=2.2.0,<2.3.0", # opencv maximum
    "opencv-python-headless>=4.12.0,<5.0.0",
    "rich>=14.1.0",
    "typer-slim>=0.19.1,<0.20.0", # pre-1.x
]

[project.scripts]
gradupe = "gradupe.__main__:cli"

[project.urls]
Home = "https://github.com/wavim/gradupe"
Repository = "https://github.com/wavim/gradupe"
Documentation = "https://github.com/wavim/gradupe#readme"
Issues = "https://github.com/wavim/gradupe/issues"
